refactor: move logger and formatted_log_message to core/logging/
Separates shared logging infrastructure from the logging UI feature.
