+++ lib/core/logging/formatted_log_message.dart
