+++ lib/core/logging/logger.dart
@@ -1,5 +1,5 @@
 import 'package:logger/logger.dart';
-import 'package:paperless_mobile/features/logging/models/formatted_log_message.dart';
+import 'package:paperless_mobile/core/logging/formatted_log_message.dart';
 
 late Logger logger;
