+++ lib/core/repository/label_repository.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:paperless_api/paperless_api.dart';
-import 'package:paperless_mobile/features/logging/data/logger.dart';
+import 'package:paperless_mobile/core/logging/logger.dart';
 
 class LabelRepository extends ChangeNotifier {
   final PaperlessLabelsApi _api;
